Remove Metarequest gem and add byebug gem.
When files were uploaded with Paperclip an exception for undefined
conversion "\xFF" from ASCII-8BIT to UTF-8 was thrown.  Historically,
Metarequest was used little to develop this project, and so it can be
removed without any ill effects.  The byebug gem and the console
debugger was used to track down this interaction, so byebug is added to
the development gem group.