Add content type validation to blog_image model.
Add content type validation to blog_image model.  This is required by
recent versions of the Paperclip gem.  It assures that the file which is
uploaded is of the same type that is expected.  Useful to help prevent
an attacker from uploading malicious files which can damage any clients
who may download them.