Improve routes organization and url consistency.
Move the unpublished blog post route into the resources blok for blog
post routes.  Change the path used for blog post slugs to be more
consistent with the other blog post paths that are generated from the
resources block.  For example now all of the paths for blog posts are
in the form:  /blog_posts/something/maybesomethingelse.