Fix issue in blog_post.set_created_at_month_and_year.
Fix blog_post.set_created_at_month_and_year to update the month and year
correctly if it is called from both a new and existing blog post.  The
previous implementation would only work for existing posts, as if you
had a new BlogPost object, the updated_at would not yet be set.  Check
if updated_at is nil and if it is then use the current date to populate
the fields else use the date store in updated_at.